feat: ship Folia extraction gameplay, quest NPC AI, and extract/hub hardening
Persist raid loot to vault on extract with hub return first, strip vanilla villager AI from quest NPCs, and land the broader extraction/lobby/proxy systems built around that.

+++ servers/paper-core/src/main/java/network/skypvp/paper/model/WorldPresetMeta.java
@@ -1,4 +1,12 @@
 package network.skypvp.paper.model;
 
-public record WorldPresetMeta(String presetId, String description, int spawnX, int spawnY, int spawnZ, float spawnYaw, float spawnPitch) {
+public record WorldPresetMeta(
+        String presetId,
+        String description,
+        double spawnX,
+        double spawnY,
+        double spawnZ,
+        float spawnYaw,
+        float spawnPitch
+) {
 }
